pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "artdetectors"
version = "0.1.0"
description = "CLIP, BLIP, SuSy, and SDXL-based image analysis + restyling pipeline"
authors = [
  { name = "Your Name" }
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  # core
  "torch",            # install with MPS build on Apple Silicon
  "torchvision",
  "Pillow",
  "transformers",
  "huggingface_hub",
  "ftfy",
  "regex",
  "tqdm",

  # image restyling (SDXL img2img)
  "diffusers>=0.31.0",
  "accelerate>=0.30.0",
  "safetensors>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/yourname/artdetectors"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"artdetectors" = [
  "data/style_clip_features.pt",
  "data/style.txt",
]

